refactor(client): adapt imports + fix UTs
