+++ src/client/components/QuickStatus/index.ts
@@ -1,1 +1,2 @@
 export { default } from './QuickStatus';
+export type { QuickStatusProps } from './QuickStatus';
