+++ src/client/components/Teleinfo/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Teleinfo';
+export type { NotificationData } from './Teleinfo';
